Enhance user and project management functionality
- Updated `UsersHandlers` to include user repository and localization support, improving user existence checks and error messaging.
- Simplified error handling in `ManageUserService` for non-existent users.
- Added new methods in `ProjectHandler` for project retrieval and updates.
- Updated `IProjectHandler` and `IProjectService` interfaces to reflect new project update capabilities.
- Implemented project status and information updates in `ProjectService` with appropriate role checks.
- Modified API routes in `IdentityController` for versioning.
- Added new endpoint in `ProjectsController` for updating project information.
- Introduced `UpdateProjectInfo` class to structure project update details.

+++ Identity/Identity.Application/Handlers/Handlers/UsersHandlers.cs
@@ -1,17 +1,21 @@
 ﻿using Identity.Identity.Application.Handlers.IHandlers;
 using Identity.Identity.Domain.Services.IServices;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Localization;
+using Module.Identity.Domain.IRepositories;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using TaskManager.SharedLayer.Localizer;
 using TaskManager.SharedLayer.RequestModels.Identity;
 using TaskManager.SharedLayer.ResponseModel;
 using TaskManager.SharedLayer.ResponseModels;
 
 namespace Identity.Identity.Application.Handlers.Handlers
 {
-    public class UsersHandlers(IManageUserService _userService) : IUsersHandlers
+    public class UsersHandlers(IManageUserService _userService, IUserRepository _userRepo, IStringLocalizer<SharedResource> _localizer) : IUsersHandlers
     {
         public async Task<ResponseModel<bool>> AddUserAsync(AddNewUserDTO model)
         {
@@ -22,6 +26,13 @@ namespace Identity.Identity.Application.Handlers.Handlers
 
         public async Task<ResponseModel<bool>> DeleteUserAsync(int Id, UpdateUserStatus model)
         {
+            var userId = Id;
+            var user = await _userRepo.GetById(userId, null, false);
+
+
+            if (user == null)
+                return new ResponseModel<bool> { Success = false, Message = _localizer["UserNotFound"] };
+
             var users = await _userService.DeleteUser(Id, model);
 
             return users;
@@ -37,6 +48,12 @@ namespace Identity.Identity.Application.Handlers.Handlers
 
         public async Task<ResponseModel<bool>> UpdateUserStatusAsync(int Id, UpdateUserStatus model)
         {
+
+            var userId = Id;
+            var user = await _userRepo.GetById(userId, null, false);
+
+
+
             var users = await _userService.UpdateUserStatus(Id, model);
 
             return users;
